Melange shouldn't even *think* about sticking "$anonymous-N" tags onto the items in enumerations anymore. It now honors Peter's new "anonymous?" flag and leaves off the enumeration name if the enumeration doesn't have one.
This is a careful but localized fix. I don't understand how the container
system works when calling compute-dylan-name, and I'm not entirely
convinced that <enum-declaration>s might not still be broken in some other
fashion. But this is clearly no worse than the current situation.

I will be investigating further. And why don't <enum-declarations> work
like structs and unions, anyway? =)