This is a workaround for the bug discovered by Neel Krishnaswami, where a variable named cluster_0_top is referenced, but never defined.
The problem was that the mrimitive values-sequence uses a stack cluster,
but analyze assumes that all operations (<operation> is the superclass
of <primitive>) won't consume a stack cluster.

This fix adds a method on <primitive> to the g.f. analyze. This is just
a workaround, as it now assumes that *all* primitive operations need
a stack cluster. This costs us 4 bytes per cluster on the stack.

What needs to be done is analysis of the primitive, to see whether
it needs a cluster or not. The design of the primitive handling
is in our way here: all the different primitives are of the same type,
and just differentiated by a slot primitive-name, which we need to
look at.