Compare <single-float> to <single-float>.
This fixes most, but not all, of the errors.  There seems to still be an
issue with precision and comparisons.  "almost-equal" checks out many
decimals past the available precision of a single-float, so it's probably
comparing random garbage.

To fix that, I'll have to overload "almost-equal", which is probably a
good idea.