[interpreter] Use _setjmp/_longjmp.
Due to an error in include order, we were using setjmp/longjmp which
are very slow on some platforms due to interacting with sigprocmask.

Now, we use _setjmp/_longjmp in the code and rely upon the config.h
to let platforms without those to work. (Are there any?)

This new organization of the code is more resistant to making mistakes
in order of includes, etc.

* mindy/config.h.cmake: Deal with missing _setjmp/_longjmp by defining
  them to be setjmp/longjmp.

* mindy/interpreter/debug.c
  (blow_off_cmd): Use _longjmp rather than longjmp.
  (invoke_debugger): Use _setjmp rather than setjmp.

* mindy/interpreter/driver.c: Move #ifndef'ery from here to config.h.cmake.